+++ config/deploy/stage.rb
@@ -3,6 +3,6 @@
 set :application, 'fleetyards-stage'
 set :rails_env, 'staging'
 set :deploy_to, '/home/fleetyards-stage'
-set :branch, 'master'
+set :branch, 'releases'
 
 server 'erebor.mortik.xyz', user: 'fleetyards-stage', roles: %w[web app db migration]